Fix running operator image tests with prebuilt image
Removed unneeded map from ClusterOperatorTest that sets the ImagePullPolicy bc. this was unexpected.

Now we're using `IfNotPresent` (as it was in v1 of this PR in the map) but from the expected minikube.yml and as such we have one source of truth (the yml files) and nothing intercepting and manipulating it in the code.

Tested locally using the example I added to the readme (using operator nightly build).

Closes #9898

Co-authored-by: Andrea Peruffo <andrea.peruffo1982@gmail.com>